+++ extract_msg/message_base.py
@@ -44,7 +44,7 @@ class MessageBase(MSGFile):
         :param overrideEncoding: optional, an encoding to use instead of the one
             specified by the msg file. Do not report encoding errors caused by
             this.
-        :param attachmentErrorBehavior: Optional, the behaviour to use in the
+        :param attachmentErrorBehavior: Optional, the behavior to use in the
             event of an error when parsing the attachments.
         :param recipientSeparator: Optional, Separator string to use between
             recipients.